refactor: MAT-176 시간 관련 유틸 함수들을 timeUtils로 공통화
